refactor: Moved AnnotationTab to subdirectory
